requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
SQLAlchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Templates
Jinja2==3.1.2

# File handling
aiofiles==23.2.1
Pillow==10.1.0

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2